requirements.txt
# ==========================================
# V17 Trading Engine - requirements.txt
# ==========================================

# Core Data Science & Matrix Operations
pandas>=2.0.0
numpy>=1.24.0

# Market Data Ingestion (High-Frequency & Macro)
alpaca-py>=0.11.0
yfinance>=0.2.30

# Quantitative Evaluation & Visualization
alphalens-reloaded>=0.4.3
matplotlib>=3.7.0